Aula 06.09 - Tratamento de erro 404

+++ src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -2,6 +2,7 @@ package br.com.alura.forum.controller;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import javax.transaction.Transactional;
 import javax.validation.Valid;
@@ -61,25 +62,45 @@ public class TopicosController {
 	}
 	
 	@GetMapping("/{id}")
-	public DetalhesDoTopicoDTO detalhar(@PathVariable Long id) {
+	public ResponseEntity<DetalhesDoTopicoDTO> detalhar(@PathVariable Long id) {
 		
-		Topico topico = topicoRepository.getOne(id);
-		return new DetalhesDoTopicoDTO(topico);
+		Optional<Topico> topico = topicoRepository.findById(id);
+		
+		if(topico.isPresent()) {
+			
+			return ResponseEntity.ok(new DetalhesDoTopicoDTO(topico.get()));
+		}
+		
+		return ResponseEntity.notFound().build();
 	}
 	
 	@PutMapping("/{id}")
 	@Transactional
 	public ResponseEntity<TopicoDTO> atualizar (@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicoForm form){
 		
-		Topico topico = form.atualizar(id, topicoRepository);
-		return ResponseEntity.ok(new TopicoDTO(topico));
+		Optional<Topico> optional = topicoRepository.findById(id);
+		
+		if(optional.isPresent()) {
+			
+			Topico topico = form.atualizar(id, topicoRepository);
+			return ResponseEntity.ok(new TopicoDTO(topico));			
+		}
+		
+		return ResponseEntity.notFound().build();
 	}
 	
 	@DeleteMapping("/{id}")
 	@Transactional
 	public ResponseEntity<?> remover (@PathVariable Long id){
 		
-		topicoRepository.deleteById(id);
-		return ResponseEntity.ok().build();
+		Optional<Topico> optional = topicoRepository.findById(id);
+		
+		if(optional.isPresent()) {
+			
+			topicoRepository.deleteById(id);
+			return ResponseEntity.ok().build();		
+		}
+
+		return ResponseEntity.notFound().build();
 	}
 }
